requirements.txt
# Core
numpy>=1.24.0
scipy>=1.10.0
numba>=0.57.0
cosysairsim>=1.8.1
jax>=0.9.1,<0.10
jaxlib>=0.9.1,<0.10


# Perception / planning (used once those modules are implemented)
opencv-python>=4.8.0
pyyaml>=6.0
matplotlib>=3.7.0
pandas>=2.0.0

# Dev
pytest>=7.4.0
black>=23.0.0